requirements.txt
antlr4-python3-runtime==4.9.3
audioread==3.0.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
cloudpickle==3.1.0
colorama==0.4.6
decorator==5.1.1
demucs==4.0.1
dora_search==0.1.12
einops==0.8.0
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
fsspec==2024.12.0
future==1.0.0
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
julius==0.2.7
lameenc==1.7.0
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
omegaconf==2.3.0
openai-whisper==20240930
openunmix==1.3.0
packaging==24.2
pillow==10.2.0
platformdirs==4.3.6
pooch==1.8.2
psutil==6.1.1
pycparser==2.22
pydub==0.25.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
scikit-learn==1.6.0
scipy==1.14.1
setuptools==75.6.0
six==1.17.0
soundfile==0.12.1
soxr==0.5.0.post1
submitit==1.5.2
sympy==1.13.1
threadpoolctl==3.5.0
tiktoken==0.8.0
torch==2.5.1
torchaudio==2.5.1+cpu
torchvision==0.20.1+cpu
tqdm==4.67.1
treetable==0.2.5
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1